debug

+++ ft_substr.c
@@ -19,12 +19,12 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	if (!s)
 		return (0);
 	if (ft_strlen(s) < start)
-		(unsigned int) len = 0;
+		len = 0;
 	if ( ft_strlen(s + start) < len)
-		(unsigned int) len = ft_strlen(s + start);
-	str = (char *) malloc (sizeof(char) * ((unsigned int)(len + 1)));
+		len = ft_strlen(s + start);
+	str = (char *) malloc (sizeof(char) * (len + 1));
 	if (!str)
 		return (0);
-	ft_strlcpy(str, s + start, (unsigned int)len + 1);
+	ft_strlcpy(str, s + start, len + 1);
 	return (str);
 }
